DjangoWebApplication/requirements.txt
pandas==1.1.2
Django==3.0.7
django_multiselectfield==0.1.12
scikit_surprise==1.1.1
Pillow==8.2.0
scikit_learn==0.24.1
surprise==0.1
